Fixed a bug that happened when changing the format, resolution or FPS of the camera.
When you changed the format, resolution or FPS, Webcamoid was turning on
and off the webcam quickly many times, if you had a webcam with a led
indicator, you can notice the led was blinking many times when you
changed the format.
I think this is related to issue #148, this bug was probably trigggering
a hardware bug that did not allowed the camera to recover its normal
parameters.