Icons: add Paperclip icons for feature snapshot to clipboard
Adds paperclip icons to the theme. Sourced from openclipart:

https://openclipart.org/detail/250668/app-icons

Note: command for semi-automation of updating icons.qrc:

sed -E '/no\./ {p;
s,(.*hicolor)/(.*)/(.*)(\.(png|svg)(.*)),\1/\2/paperclip\4, }'
StandAlone/icons.qrc

This matches on the LINE BEFORE where the new line insert is required (in this
case on the "no.png" line), prints that line, then edits it to replace
the filename with "paperclip" and prints the modifed line. Result is an appended
line with the new filename with the correct path for the size of the
icon.