Fix: ensure pkexec can always execute shell scripts
Debian, Ubuntu and deriviatives use a forked version of the polkit
package (that contains "pkexec"). When no command line is passed it
causes pkexec to fail (with its --help).

This causes installation, removal, and configuration of Virtual Cameras
to fail.

The fork occured in 2013 when upstream polkit replaced the configuration
system with executable Javascript using the Mozilla SpiderMonkey engine.

Upstream later added functionality that will get the user's shell from
/etc/passwd if the pkexec command line is empty.

That allowed webcamoid work with upstream polkit versions but fail with
Debian and derivatives.

Solve this by always passing "sh" on the pkexec command-line.

Closes issue #510.