MultiSrc: expose sync property to GUI
This completes the work done in commit 8ce3f3442 (2020-04-11) that added the
ability to control whether multimedia sources synchronise themselves to the
presentation time stamp (PTS) or not.
The option defaults to true but for some sources such as RTSP (e.g:
network cameras) synchronising to PTS can cause large (> 2 second) latency.

Allow the operator to disable sync so that frames are processed as soon as
they arrive.

Closes: issue #85.