Plugin: add Crop effect
Crop uses slider controls to select left, top, width and height of area
to crop from the source frame. It maintains the same aspect ratio and
automatically reduces width and height when left or top positions are
increased. It sets a minimum width and height at 10% of the source.

This addresses one part of issue #72.