Set MACOSX_DEPLOYMENT_TARGET in macOS CI
On macOS each build must specify for which OS version the binary should be
built (i.e. the earliest OS version which should be supported).
If this isn't done, CMake "compute[s] [it] based on the host platform".
https://cmake.org/cmake/help/v3.24/variable/CMAKE_OSX_DEPLOYMENT_TARGET.html
The target version can be directly specified by the CMAKE_OSX_DEPLOYMENT_TARGET
CMake variable or the MACOSX_DEPLOYMENT_TARGET environment variable.
This commit does the latter.