Disabled upload of Mac packages.
As explained in issue #584, all packages for Apple Mac M1 requires that
all apps to be code signed, this requirement is imposible to comply
with, so the upload of binaries are disabled because they are simply
unusable. You can still compile Webcamoid from the source code.
Developers and maintainers for Mac are welcome.